imp - Properly calculate thresholds
---

We need to properly calculate the thresholds to ensure that we don't render an image that is clearly wider or taller than the specified dimensions.

---

Type: imp
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1

+++ Terminaux.Images/ImageProcessor.cs
@@ -131,8 +131,8 @@ namespace Terminaux.Images
             // Get the image width and height in pixels and get their comparison factor
             int imageWidth = imageColors.GetLength(0);
             int imageHeight = imageColors.GetLength(1);
-            double imageWidthThreshold = imageWidth / width;
-            double imageHeightThreshold = imageHeight / height;
+            double imageWidthThreshold = (double)imageWidth / width;
+            double imageHeightThreshold = (double)imageHeight / height;
 
             // Build the buffer
             StringBuilder buffer = new();
